Leave the deployment ssh key in the VM
With `ansible` version `2.5.3` the playbook formally fails if the SSH used by Ansible is removed.
It is still possible to maintain the previous behaviour, enabling *only* the ssh key of the user in the VM, but the Cloud Portal will report a deployment failure because Ansible is failing its last task.